Move some logic to controller
