+++ AirCasting/SearchAndFollow/CompleteScreen/Models/PartialExternalSession.swift
