+++ AirCasting/SearchAndFollow/CompleteScreen/Views/CompleteScreenViewModel.swift
@@ -56,9 +56,9 @@ class CompleteScreenViewModel: ObservableObject {
     private let session: PartialExternalSession
     private var externalSessionWithStreams: ExternalSessionWithStreamsAndMeasurements?
     
-    @Injected private var service: StreamDownloader
     @Injected private var singleSessionDownloader: SingleSessionDownloader
     @Injected private var externalSessionsStore: ExternalSessionsStore
+    @Injected private var controller: SearchAndFollowCompleteScreenController
     
     init(session: PartialExternalSession, exitRoute: @escaping () -> Void) {
         self.session = session
@@ -123,7 +123,7 @@ class CompleteScreenViewModel: ObservableObject {
     private func getMeasurementsAndDisplayData() {
         let streams = session.stream.map(\.id)
         
-        downloadMeasurements(streams: streams) { [weak self] result in
+        controller.downloadMeasurements(streams: streams) { [weak self] result in
             guard let self = self else { return }
             switch result {
             case .failure(let error):
@@ -135,7 +135,7 @@ class CompleteScreenViewModel: ObservableObject {
                 guard !downloadedStreamsWithMeasurements.isEmpty else { return }
                 
                 DispatchQueue.main.async {
-                    self.externalSessionWithStreams = self.createExternalSession(with: downloadedStreamsWithMeasurements)
+                    self.externalSessionWithStreams = self.controller.createExternalSession(from: self.session, with: downloadedStreamsWithMeasurements)
                     
                     self.sessionStreams = .ready( self.externalSessionWithStreams!.streams.map {
                         .init(id: $0.id,
@@ -156,61 +156,7 @@ class CompleteScreenViewModel: ObservableObject {
         }
     }
     
-    private func createExternalSession(with downloadedStreamsWithMeasurements: [StreamWithMeasurements]) -> ExternalSessionWithStreamsAndMeasurements {
-        .init(uuid: self.session.uuid,
-              provider: self.session.provider,
-              name: self.session.name,
-              startTime: self.session.startTime,
-              endTime: self.session.endTime,
-              longitude: self.session.longitude,
-              latitude: self.session.latitude,
-              streams: self.session.stream.compactMap { stream in
-            
-            guard let downloadedStream = downloadedStreamsWithMeasurements.first(where: { $0.sensorName == stream.sensorName }) else { return nil }
-            
-            let measurements = downloadedStream.measurements
-            return .init(id: stream.id,
-                         unitName: stream.unitName,
-                         unitSymbol: stream.unitSymbol,
-                         measurementShortType: stream.measurementShortType,
-                         measurementType: stream.measurementType,
-                         sensorName: stream.sensorName,
-                         sensorPackageName: stream.sensorPackageName,
-                         thresholdVeryLow: stream.thresholdsValues.veryLow,
-                         thresholdLow: stream.thresholdsValues.low,
-                         thresholdMedium: stream.thresholdsValues.medium,
-                         thresholdHigh: stream.thresholdsValues.high,
-                         thresholdVeryHigh: stream.thresholdsValues.veryHigh,
-                         thresholdsValues: stream.thresholdsValues,
-                         measurements: measurements.map {.init(value: $0.value, time: DateBuilder.getDateWithTimeIntervalSince1970(Double($0.time/1000)), latitude: $0.latitude, longitude: $0.longitude)})
-        })
-    }
     
-    private func downloadMeasurements(streams: [Int], completion: @escaping (Result<[StreamWithMeasurements], Error>) -> Void) {
-        guard !streams.isEmpty else {
-            completion(.failure(CompletionScreenError.noStreams))
-            return
-        }
-        var results: [Result<StreamWithMeasurements, Error>] = []
-        let measurementsLimit = 60*24 // We want measurements from 24 hours
-        let group = DispatchGroup()
-        streams.forEach { streamId in
-            group.enter()
-            self.service.downloadStreamWithMeasurements(id: streamId, measurementsLimit: measurementsLimit) { results.append($0); group.leave() }
-        }
-        group.notify(queue: .global()) {
-            var allDownstreams = [StreamWithMeasurements]()
-            for result in results {
-                do {
-                    allDownstreams.append(try result.get())
-                } catch {
-                    completion(.failure(error))
-                    return
-                }
-            }
-            completion(.success(allDownstreams))
-        }
-    }
     
     private static func getSensorName(_ streamName: String) -> String {
         streamName
